// src/enums/appEnum.ts
export enum MenuTypeEnum {
  LEFT = 'left',
  TOP = 'top',
  TOP_LEFT = 'top-left',
  DUAL_MENU = 'dual-menu'
}

export enum SystemThemeEnum {
  DARK = 'dark',
  LIGHT = 'light',
  AUTO = 'auto'
}

export enum MenuThemeEnum {
  DARK = 'dark',
  LIGHT = 'light',
  DESIGN = 'design'
}

export enum MenuWidth {
  CLOSE = '70px'
}

export enum LanguageEnum {
  ZH = 'zh',
  EN = 'en'
}

export enum IconTypeEnum {
  CLASS_NAME = 'className',
  UNICODE = 'unicode'
}

export enum ContainerWidthEnum {
  FULL = '100%',
  BOXED = '1200px'
}

export enum BgColorEnum {
  PRIMARY = 'bg-primary',
  SECONDARY = 'bg-secondary',
  WARNING = 'bg-warning',
  ERROR = 'bg-error',
  SUCCESS = 'bg-success',
  DANGER = 'bg-danger',
  INFO = 'bg-info'
}